% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_clust.R
\name{assign_clust}
\alias{assign_clust}
\title{Function to assign population to centers of services}
\usage{
assign_clust(clustered)
}
\arguments{
\item{clustered}{List of the population assigned to each centre by EPP::clust_it}
}
\value{
list with Centres of the clusters with the number of individuals covered for each service distance and mean distances and Population with the center assigned to
}
\description{
Function to process the results from EPP::clust_it
}
\examples{
a <- clust_it(pop_epp)
a <- assign_clust
}
