% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centers_epp}
\alias{centers_epp}
\title{Simulated Centers of attention in Uruguay.}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{x}{x coords in EPSG:31721 CRS}
  \item{y}{y coords in EPSG:31721 CRS}
  \item{id}{id of the center}
  \item{capacity}{capacity of the centers}
}
}
\usage{
centers_epp
}
\description{
A dataset containing the coords and other attributes of 10 centers of attention.
}
\keyword{datasets}
