% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-epp-files.R
\name{.parse_array}
\alias{.parse_array}
\title{Parse java array from EPP .xml file}
\usage{
.parse_array(x)
}
\arguments{
\item{x}{`xml_node` object with tag 'array'}
}
\description{
Parse java array from EPP .xml file
}
