% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-epp-files.R
\name{.parse_matrix}
\alias{.parse_matrix}
\title{Parse java matrix from EPP .xml file}
\usage{
.parse_matrix(x)
}
\arguments{
\item{x}{`xml_node` object with tag 'array' and type '[D' or '[I'}
}
\description{
Parse java matrix from EPP .xml file
}
