% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epp.R
\docType{package}
\name{epp}
\alias{epp}
\alias{pop}
\title{\code{epp} package}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
pop
}
\description{
Evaluation of Proximity Programs
}
\details{
See the README on
\href{https://gitlab.mides.gub.uy/rdetomasi/epp/blob/master/README.md}{Gitlab}
}
\keyword{datasets}
