% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-epp-files.R
\name{get_eppxml_workset}
\alias{get_eppxml_workset}
\title{Return workset with names from EPP .xml file}
\usage{
get_eppxml_workset(pjnz)
}
\arguments{
\item{x}{`xml_node` object with tag 'array'}
}
\description{
Return workset with names from EPP .xml file
}
