% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labels_es}
\alias{labels_es}
\title{Labels to translate leafepp maps.}
\format{
A data frame with 16 rows and 2 variables:
\describe{
  \item{en}{English labels}
  \item{es}{Spanish labels}
}
}
\usage{
labels_es
}
\description{
A dataset containing the labels to translate leafepp maps.
}
\keyword{datasets}
