% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafepp.R
\name{leafepp}
\alias{leafepp}
\title{leafepp}
\usage{
leafepp(x, type, crs, ...)
}
\arguments{
\item{x}{List obtained from eppexist or eppproy}

\item{type}{Character "exist" or "proy" depending of input structure}

\item{crs}{Coordinate Reference Systems (CRS).}

\item{...}{leaflet options and parameters}
}
\value{
leaflet
}
\description{
leafepp
}
\examples{
## In case of eppexist

exist <- eppexist(pop = pop_epp, 
                  centers = centers_epp, 
                  crs = sp::CRS("+init=epsg:32721"))
leafepp(exist, type = "exist", crs = sp::CRS("+init=epsg:32721"))

## In case of eppproy

proy <- eppproy(pop = exist$pop_uncover)
leafepp(proy, type = "proy", crs = sp::CRS("+init=epsg:32721"))
}
