% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_epp}
\alias{pop_epp}
\title{Simulated Population in Uruguay.}
\format{
A data frame with 500 rows and 3 variables:
\describe{
  \item{x}{x coords in EPSG:31721 CRS}
  \item{y}{y coords in EPSG:31721 CRS}
  \item{weight}{weight variable to analyze}
}
}
\usage{
pop_epp
}
\description{
A dataset containing the coords and other attributes of 500 individuals.
}
\keyword{datasets}
