% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{prepare_ancsite_likdat}
\alias{prepare_ancsite_likdat}
\title{Prepare site-level ANC prevalence data for EPP random-effects likelihood}
\usage{
prepare_ancsite_likdat(eppd, anchor.year = 1970L)
}
\arguments{
\item{eppd}{EPP data object}

\item{anchor.year}{year in which EPP data inputs start
NOTE: requires year to be stored in column names of anc.prev}
}
\description{
Prepare site-level ANC prevalence data for EPP random-effects likelihood
}
