% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-epp-files.R
\name{read_epp_data}
\alias{read_epp_data}
\title{Read EPP fitting surveillance data}
\usage{
read_epp_data(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file.}
}
\description{
Reads the HIV prevalence from sentinel surveillance and household surveillance
data from the EPP .xml file within a PJNZ.
}
\details{
EPP projection sets are identified in the .xml file by searching the XML tree
for tag "object", and then selecting objects with "class" attribute equal to
"epp2011.core.sets.ProjectionSet".
}
