% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-epp-files.R
\name{read_epp_subpops}
\alias{read_epp_subpops}
\title{Read EPP subpopulation configuration}
\usage{
read_epp_subpops(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file.}
}
\description{
Reads the subpopulation configuration and population sizes from the EPP .xml
file within a PJNZ.
}
\details{
EPP projection sets are identified in the .xml file by searching the XML tree
for tag "object", and then selecting objects with "class" attribute equal to
"epp2011.core.sets.ProjectionSet".
}
