% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-epp-files.R
\name{read_epp_workset_options}
\alias{read_epp_workset_options}
\title{Read EPP Workset Options}
\usage{
read_epp_workset_options(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file}
}
\description{
Parse 'Worset option' selections for overall EPP
workset in the "Model options" pop up box in the
Curve Fitting workflow
}
