% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-epp-files.R
\name{read_eppxml_results}
\alias{read_eppxml_results}
\title{Read Bayesian Median outputs from XML}
\usage{
read_eppxml_results(pjnz)
}
\arguments{
\item{pjnz}{}
}
\description{
Read Bayesian Median outputs from XML
}
